Fast assembly conversion to hex string
On a test database, OpenDBDiff is taking 27% of its runtime in GenerateAssemblies:Fill. Most of this time is taken to convert the assembly bytes to a hex string. Here this code is optimized with a lookup technique. After the change, GenerateAssemblies:Fill only takes 3% of the total runtime.

Reference for hex encoding performance: https://stackoverflow.com/questions/311165/how-do-you-convert-a-byte-array-to-a-hexadecimal-string-and-vice-versa