Use Tsql-Parser to format the CREATE calls instead of Regex
The code in FormatCode is using several Regex objects to parse and normalize the CREATE statements. This tends to be slow, mostly because Regex objects are usually better when they are static. However, generalizing the Regex used here to make them static also makes them much slower.

Instead of trying to fix the Regex, here this code is replaced by the use of the Tsql-parser library (an open source project on Github). This is much faster, and also simplifies the code, since the complex Regex used to detect SQL comments are not needed anymore.

Before the change, FormatCode:FindCreate takes 17.9 % of the runtime on a test database.
After the change, FormatCode:FindAndNormalizeCreate (the equivalent method, renamed) is only taking 6.3 %.